+++ CHANGELOG.md
@@ -558,6 +558,7 @@
 ## [unreleased]
 
 - Users can report the strategic area under which the GCRF allocation was made
+- Bring back script for importing forecast data into an in-review/approved report
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-39...HEAD
 [release-39]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-38...release-39